[stable/kong] specify ingress class for Ingress Controller (#10031)
If multiple Kong Ingress controllers are deployed with different
ingress-class parameters to satisfy different kinds of ingresses
in the same k8s cluster,
they will attempt to take aquire leader status on the same
election key, meaning only one of the ingress-class will ever be
satisfied.

This change ensures that a unique key is used for leader election for
every ingress-class.

Signed-off-by: Harry Bagdi <harrybagdi@gmail.com>