[stable/rabbitmq-ha] Fixed invalid value (#18503)
* [stable/rabbitmq-ha] Fixed invalid value

Set the service.clusterIP variable to an empty value and commented out,
following the standard set in other charts for optional empty values, to
avoid invalid value error. Also updated readme and bumped chart version.

Signed-off-by: Austin Orth <aorth@niche.com>

* Implemented changes suggested by reviewers

Reverted to the previous settings for the value and instead updated the
logic in the service template to avoid setting clusterIP value when type
is LoadBalancer and it is set to "None".

Signed-off-by: Austin Orth <aorth@niche.com>